Update base.py in order to fix web2py-issues/697
This is an addition to the prior fix that introduces a problem with the incoming list objects are already strings. By wrapping the code in a try/except where looking specifically for a UnicodeEncodeError we maintain the existing code in all cases except when an actual error is generated. This should reduce the possibility of carry-on problems.