MongoDB ondelete support for list:reference
CASCADE and SET NULL support for list:reference fields.  For CASCADE
field if the reference is the only item in the reference list then
remove the entire record 'ala' non-list based references.  For CASCADE
fields with multiple references and for SET NULL fields, simply remove
the reference id from the list.