fixes a few intricacies properly escaping LIKE statements
specifically, we used like with startswith and endswith, and contains,
improperly. We need correct (and "automagic") escaping when dealing with
those "pythonic" methods, else subtle bugs creep in.
As always, "like()" and "ilike()" maps to the exact statement in the
backend,with the usual behaviour.
like() and ilike() got also a new "escape" parameter in order to pass
whatever escaping character the user feel right. By default, the escape
character is `\`, and the term passed to the like is escaped with it.

well, it's been fun. 1st absolute time that a feature gets backported
from mssql to all t-sql backends. As usual, NoSQL is out of my reach,
and left as an exercise to the ones more well-versed into the matter.