Set version number to 19.02 to make sure it's released this month
Changed version number to reduce missteps. It should be clear to developers what the current version is. Currently it's nowhere to be found, not in setup.py (which uses this version number), not in CHANGES (which is outdated) and not in README.md (which states that version 17.11 is the latest version on PyPi.