Enabled db(query, ignore_common_filters=True) for IS_IN_DB
The IS_IN_DB validator was ignoring the ignore_common_filters option when it was added to the Set definition. It now checks if this has been specified. This improves the previous change which only checked if the ignore_common_filters option was specified in the IS_IN_DB function.